pyproject.toml
[project]
name = "pkglink"
description = "Create symlinks to python package directories"
authors = [{ name = "hotdog-werx" }]
readme = "README.md"
license = 'MIT'
requires-python = ">=3.11"
dynamic = ["version"]

dependencies = [
  "hotlog>=0.1.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0, <7",
  "pathspec>=0.12.0, <1",
  "structlog>=25.4.0, <26",
  "rich>=14.0.0, <15",
  "argcomplete>=3.6.2, <4",
  "tomli-w>=1.2.0",
]

[project.scripts]
plk = "pkglink.cli.plk:main"
pkglink = "pkglink.cli.pkglink:main"
pkglinkx = "pkglink.cli.pkglinkx:main"
pkglink-batch = "pkglink.cli.pkglink_batch:main"

[dependency-groups]
dev = [
  "tbelt>=0.0.3",
  "pytest>=8.0.0",
  "pytest-mock>=3.12.0",
  "pytest-cov>=4.0.0",
]

[tool.hatch]
build.packages = ["pkglink"]
metadata.allow-direct-references = true
version.path = 'pkglink/version.py'

[tool]
pyright.extends = ".codeguide/configs/pyrightconfig.json"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
